+++ src/titan/ODESolverInterface.java
@@ -1,0 +1,52 @@
+/*
+ * @author Pieter Collins, Christof Seiler, Katerina Stankova, Nico Roos, Katharina Schueller
+ * @version 0.99.0
+ *
+ * This interface serves as the API for students in phase 1.
+ */
+
+package titan;
+
+import titan.StateInterface;
+import titan.ODEFunctionInterface;
+
+/*
+ * A class for solving a general differential equation dy/dt = f(t,y)
+ *     y(t) describes the state of the system at time t
+ *     f(t,y(t)) defines the derivative of y(t) with respect to time t
+ */
+public interface ODESolverInterface {
+
+    /*
+     * Solve the differential equation by taking multiple steps.
+     *
+     * @param   f       the function defining the differential equation dy/dt=f(t,y)
+     * @param   y0      the starting state
+     * @param   ts      the times at which the states should be output, with ts[0] being the initial time
+     * @return  an array of size ts.length with all intermediate states along the path
+     */
+    public StateInterface[] solve(ODEFunctionInterface f, StateInterface y0, double[] ts);
+
+    /*
+     * Solve the differential equation by taking multiple steps of equal size, starting at time 0.
+     * The final step may have a smaller size, if the step-size does not exactly divide the solution time range
+     *
+     * @param   f       the function defining the differential equation dy/dt=f(t,y)
+     * @param   y0      the starting state
+     * @param   tf      the final time
+     * @param   h       the size of step to be taken
+     * @return  an array of size round(tf/h)+1 including all intermediate states along the path
+     */
+    public StateInterface[] solve(ODEFunctionInterface f, StateInterface y0, double tf, double h);
+
+    /*
+     * Update rule for one step.
+     *
+     * @param   f   the function defining the differential equation dy/dt=f(t,y)
+     * @param   t   the time
+     * @param   y   the state
+     * @param   h   the step size
+     * @return  the new state after taking one step
+     */
+    public StateInterface step(ODEFunctionInterface f, double t, StateInterface y, double h);
+}
